# apps/backend/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AgentBudgets": {
        "description": "Optional user-selected product budgets. Unset by default (AGT-003).",
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "max_tool_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tool Calls"
          }
        },
        "title": "AgentBudgets",
        "type": "object"
      },
...
      "AgentStartRequest": {
        "additionalProperties": false,
        "properties": {
          "budgets": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentBudgets"
              },
              {
                "type": "null"
              }
            ]
          },
          "content_blocks": {
            "anyOf": [
              {
                "items": {
                  "discriminator": {
                    "mapping": {
                      "image_url": "#/components/schemas/ImageContentBlock",
                      "text": "#/components/schemas/TextContentBlock"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TextContentBlock"
                    },
                    {
                      "$ref": "#/components/schemas/ImageContentBlock"
                    }
                  ]
                },
                "maxItems": 32,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Blocks"
          },
          "criteria": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskCriteria"
              },
              {
                "type": "null"
              }
            ]
          },
          "deployment_id": {
            "title": "Deployment Id",
            "type": "string"
          },
          "embedding_deployment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Deployment Id"
          },
          "inherit_deployment_settings": {
            "default": true,
            "title": "Inherit Deployment Settings",
            "type": "boolean"
          },
          "input_message_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Message Id"
          },
          "knowledge_version_refs": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Version Refs",
            "type": "array"
          },
          "memory_version_refs": {
            "items": {
              "type": "string"
            },
            "title": "Memory Version Refs",
            "type": "array"
          },
          "output_schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputSchemaRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "per_request_overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Request Overrides"
          },
          "presented_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presented Tools"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "project_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Path"
          },
          "protected_instruction_version_refs": {
            "items": {
              "type": "string"
            },
            "title": "Protected Instruction Version Refs",
            "type": "array"
          },
          "recorded_fixtures": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded Fixtures"
          },
          "resume_checkpoint_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Checkpoint Id"
          },
          "retrieval_project_paths": {
            "items": {
              "type": "string"
            },
            "title": "Retrieval Project Paths",
            "type": "array"
          },
          "skill_version_refs": {
            "items": {
              "type": "string"
            },
            "title": "Skill Version Refs",
            "type": "array"
          },
          "source_surface": {
            "default": "agent-run",
            "enum": [
              "agent-run",
              "chat",
              "lab"
            ],
            "title": "Source Surface",
            "type": "string"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "task": {
            "title": "Task",
            "type": "string"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "tool_mode": {
            "$ref": "#/components/schemas/ToolMode",
            "default": "live-tool"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "deployment_id",
          "task"
        ],
        "title": "AgentStartRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageContent": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "default": "auto",
            "enum": [
              "auto",
              "low",
              "high"
            ],
            "title": "Detail",
            "type": "string"
          },
          "url": {
            "maxLength": 12000000,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ImageContent",
        "type": "object"
      },
      "ImageContentBlock": {
        "additionalProperties": false,
        "properties": {
          "image_url": {
            "$ref": "#/components/schemas/ImageContent"
          },
          "type": {
            "const": "image_url",
            "default": "image_url",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "image_url"
        ],
        "title": "ImageContentBlock",
        "type": "object"
      },
...
      "InterruptDecision": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "scope": {
            "default": "once",
            "enum": [
              "once",
              "session",
              "always"
            ],
            "title": "Scope",
            "type": "string"
          },
          "type": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "InterruptDecision",
        "type": "object"
      },
      "InterruptDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/InterruptDecision"
            },
            "title": "Decisions",
            "type": "array"
          },
          "interrupt_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Id"
          },
          "namespace": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Namespace",
            "type": "array"
          }
        },
        "required": [
          "decisions"
        ],
        "title": "InterruptDecisionRequest",
        "type": "object"
      },
...
      "OutputSchemaRequest": {
        "additionalProperties": false,
        "description": "Caller-supplied JSON schema for one run.\n\nThe schema is passed to Deep Agents' existing ``response_format`` seam.\nIt is not parsed out of assistant prose.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[A-Za-z][A-Za-z0-9_]{0,119}$",
            "title": "Name",
            "type": "string"
          },
          "schema": {
            "additionalProperties": true,
            "title": "Schema",
            "type": "object"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "strategy": {
            "default": "auto",
            "enum": [
              "auto",
              "native",
              "tool"
            ],
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "OutputSchemaRequest",
        "type": "object"
      },
...
      "TaskCriteria": {
        "properties": {
          "checks": {
            "items": {
              "type": "string"
            },
            "title": "Checks",
            "type": "array"
          },
          "expected_artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Expected Artifacts",
            "type": "array"
          },
          "review_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Prompt"
          }
        },
        "title": "TaskCriteria",
        "type": "object"
      },
      "TextContentBlock": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "text",
            "default": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextContentBlock",
        "type": "object"
      },
      "ToolMode": {
        "enum": [
          "live-tool",
          "recorded-tool"
        ],
        "title": "ToolMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WorkbenchLocalToken": {
        "description": "Same-machine desktop\u2194backend shared-secret token. Electron main injects this header (Issue #40).",
        "in": "header",
        "name": "X-Workbench-Local-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Canonical shared HTTP contract surface generated from Pydantic models. Not the product route table. Product OpenAPI stays unpublished.",
    "title": "Local AI Workbench shared contracts",
    "version": "0.1.0",
    "x-workbench-generated": "GENERATED by workbench_backend.contracts \u2014 do not hand-edit. Regenerate with the registered shared-contract command."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent-runs": {
      "get": {
        "operationId": "list_agent_runs_v1_agent_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Agent Runs V1 Agent Runs Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Agent Runs"
      },
      "post": {
        "operationId": "start_agent_run_v1_agent_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Start Agent Run V1 Agent Runs Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Agent Run"
      }
    },
    "/v1/agent-runs/{run_id}": {
      "get": {
        "operationId": "get_agent_run_v1_agent_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Agent Run V1 Agent Runs  Run Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent Run"
      }
    },
    "/v1/agent-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_agent_run_v1_agent_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Agent Run V1 Agent Runs  Run Id  Cancel Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Agent Run"
      }
    },
    "/v1/agent-runs/{run_id}/interrupt-decision": {
      "post": {
        "operationId": "decide_agent_run_interrupt_v1_agent_runs__run_id__interrupt_decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterruptDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Decide Agent Run Interrupt V1 Agent Runs  Run Id  Interrupt Decision Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Agent Run Interrupt"
      }
    },
...
  }
}